requirements.txt
flask
python-dotenv
google-cloud-vertexai
google-cloud-aiplatform
google-cloud-bigquery
google-auth
elevenlabs
confluent-kafka
ddtrace
datadog
gunicorn
requests